cleanup some init
Change location of the php_version during install so we can reference the “main” PHP version

+++ internal/initialize/initialize.go
@@ -23,13 +23,13 @@ write_files:
 
       # run the correct script depending on the version of PHP
       if [ "$phpversion" == '7.3' ]; then
-        echo "7.3" > /home/ubuntu/.php_version
+        sudo echo "7.3" > /opt/nitro/php_version
         sudo bash /opt/nitro/php/php-73.sh
       elif [ "$phpversion" == '7.2' ]; then
-        echo "7.2" > /home/ubuntu/.php_version
+        sudo echo "7.2" > /opt/nitro/php_version
         sudo bash /opt/nitro/php/php-72.sh
       else
-        echo "7.4" > /home/ubuntu/.php_version
+        sudo echo "7.4" > /opt/nitro/php_version
         sudo bash /opt/nitro/php/php-74.sh
       fi
 
@@ -52,7 +52,9 @@ write_files:
       sudo bash /opt/nitro/nginx/install.sh
 
       # remove the default site
-      sudo rm /etc/nginx/sites-enabled/default
+      if [ -f /etc/nginx/sites-enabled/default ]; then
+          sudo rm /etc/nginx/sites-enabled/default
+      fi
     permissions: '770'
   - path: /opt/nitro/update.sh
     content: |
@@ -64,17 +66,17 @@ write_files:
     content: |
       #!/bin/bash
       apt install -y php7.4 php7.4-mbstring php7.4-cli php7.4-curl php7.4-fpm php7.4-gd php7.4-intl php7.4-json \
-      php7.4-mysql php7.4-opcache php7.4-pgsql php7.4-zip php7.4-xml php7.4-xdebug
+      php7.4-mysql php7.4-opcache php7.4-pgsql php7.4-zip php7.4-xml php-xdebug
   - path: /opt/nitro/php/php-73.sh
     content: |
       #!/bin/bash
       apt install -y php7.3 php7.3-mbstring php7.3-cli php7.3-curl php7.3-fpm php7.3-gd php7.3-intl php7.3-json \
-      php7.3-mysql php7.3-opcache php7.3-pgsql php7.3-zip php7.3-xml php7.3-xdebug
+      php7.3-mysql php7.3-opcache php7.3-pgsql php7.3-zip php7.3-xml php-xdebug
   - path: /opt/nitro/php/php-72.sh
     content: |
       #!/bin/bash
       apt install -y php7.2 php7.2-mbstring php7.2-cli php7.2-curl php7.2-fpm php7.2-gd php7.2-intl php7.2-json \
-      php7.2-mysql php7.2-opcache php7.2-pgsql php7.2-zip php7.2-xml php7.2-xdebug
+      php7.2-mysql php7.2-opcache php7.2-pgsql php7.2-zip php7.2-xml php-xdebug
   # create the composer install script
   - path: /opt/nitro/composer-install.sh
     content: |
